refactor: move aggregate types to own module
