+++ actix-ws/CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## Unreleased
 
-- Take the encoded buffer when yielding bytes in the response stream rather than splitting the buffer, reducing memory use
+- Take the encoded buffer when yielding bytes in the response stream rather than splitting the buffer, reducing memory use.
+- Add `AggregatedMessage[Stream]` types.
 - Remove type parameters from `Session::{text, binary}()` methods, replacing with equivalent `impl Trait` parameters.
 - `Session::text()` now receives an `impl Into<ByteString>`, making broadcasting text messages more efficient.
 - Allow sending continuations via `Session::continuation()`
-- Enable customizing `max_size` for received frames
-- Add new ability to aggregate received continuations
+- Enable customizing `max_size` for received frames.
 
 ## 0.2.5
