+++ actix-ws/src/lib.rs
@@ -15,11 +15,13 @@ use actix_http::{
 use actix_web::{web, HttpRequest, HttpResponse};
 use tokio::sync::mpsc::channel;
 
+mod aggregated;
 mod fut;
 mod session;
 
 pub use self::{
-    fut::{AggregatedMessage, AggregatedMessageStream, MessageStream, StreamingBody},
+    aggregated::{AggregatedMessage, AggregatedMessageStream},
+    fut::{MessageStream, StreamingBody},
     session::{Closed, Session},
 };
